diana-v2/backend/requirements.txt
Flask==3.0.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
pymongo==4.7.2
dnspython==2.6.1
bcrypt==4.1.3
python-dotenv==1.0.1
pytest==8.2.2
pytest-mock==3.14.0
